Renames getContacts.ts to ContactsQuery.ts
